tests/wpt/web-platform-tests/css/css-backgrounds/parsing/background-image-computed.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Backgrounds and Borders: getComputedValue().backgroundImage</title>
<link rel="help" href="https://drafts.csswg.org/css-backgrounds/#background-image">
<meta name="assert" content="background-image computed value is as specified.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
</head>
<body>
<div id="target"></div>
<script>
test_computed_value("background-image", "none");

test_computed_value("background-image", 'url("http://www.example.com/")');
test_computed_value("background-image", 'none, url("http://www.example.com/")');
</script>
</body>
</html>
